webapi/webapi-serviceworkers-w3c-tests/serviceworkers/service-worker-container-register.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>ServiceWorkerContainer.register()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="https://slightlyoff.github.io/ServiceWorker/spec/service_worker/index.html#navigator-service-worker-register">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="w3c/service-workers/resources/test-helpers.js"></script>
<script src="w3c/cache-storage/resources/testharness-helpers.js"></script>
<div id="log"></div>
<script>

promise_test(function(t) {
  var scope = "resources/"
  var scriptURL = "resources/empty-worker.js";
  return service_worker_unregister(t, scriptURL, scope)
    .then(function() {
      navigator.serviceWorker.register(scriptURL, {scope: scope})
        .then(function(registration) {
          assert_true(registration instanceof EventTarget);
          assert_true(registration instanceof ServiceWorkerRegistration);
          var absoulteScope = normalizeURL(scope);
          assert_equals(registration.scope, absoulteScope);
          return service_worker_unregister_and_done(t, scope);
        });
    });
}, "Check that ServiceWorkerContainer.register with normal scope.");

promise_test(function(t) {
  var scriptURL = "resources/empty-worker.js";
  var absoulteScript = normalizeURL(scriptURL);
  var scope = new URL("./", absoulteScript).toString();
  return service_worker_unregister(t, scriptURL, scope)
    .then(function() {
      navigator.serviceWorker.register(scriptURL)
        .then(function(registration) {
          assert_true(registration instanceof EventTarget);
          assert_true(registration instanceof ServiceWorkerRegistration);
          assert_equals(registration.scope, scope);
          return service_worker_unregister_and_done(t, scope);
        });
    });
}, "Check that ServiceWorkerContainer.register with default scope.");

promise_test(function(t) {
  var scriptURL = "resources/empty-worker.js";
  return assert_promise_rejects(
    navigator.serviceWorker.register(scriptURL, {scope: "not-exist"}),
    "SecurityError",
    "Registration scope outside the script directory should fail with SecurityError.");
}, "Check that ServiceWorkerContainer.register with not exist scope.");

promise_test(function(t) {
  var scriptURL = "http://example.com/worker.js";
  var scope = "http://example.com/scope/";
  return assert_promise_rejects(
    navigator.serviceWorker.register(scriptURL, {scope: scope}),
    "SecurityError",
    "Registration script outside domain should fail with SecurityError.");
}, "Check that ServiceWorkerContainer.register with outside domain");

</script>
